control.lua: Remove placed_explosives from old saves.
The old code added to placed_explosives but mostly did not remove
things from it, so it would grow without bound.  The new code does
not need the array, so does not create it, but if we load an old
save, we should delete the array to reclaim the space.